requirements.txt
speechrecognition
pywin32
google-generativeai
pillow
streamlit
pyaudio
